feat!: added UI support of UndirectedGraph
