+++ src/main/kotlin/view/MainScreenView.kt
@@ -5,12 +5,16 @@ import androidx.compose.runtime.Composable
 import viewmodel.MainScreenViewModel
 import androidx.compose.material.Button
 import androidx.compose.material.Text
+import model.DirectedGraph
 
 @Composable
 fun <V, E> MainScreenView(viewModel: MainScreenViewModel<V, E>) {
 	Column {
-		Button (onClick = viewModel::calculateSCC){
-			Text("Calculate SCC")
+		if (viewModel.graph is DirectedGraph) {
+			Button (onClick = viewModel::calculateSCC){
+				Text("Calculate SCC")
+			}
 		}
+
 	}
 }
